Floats numbers are now a thing lol

+++ main.py
@@ -18,40 +18,40 @@ while True: #main loop
 
     while True: 
         user_input = input(f"{YELLOW}> {BLUE}")
-        if user_input == "1":
+        if user_input == "1":           # means the user chose to calculate the time of the download
             while True:
                 try:
-                    speed = int(input(f"{YELLOW}your internet speed {RED}(Mbps){YELLOW} > {DarkBLUE}"))
-                    size = int(input(f"{YELLOW}your file size {RED}(Mbit){YELLOW} > {DarkBLUE}"))
+                    speed = float(input(f"{YELLOW}your internet speed {RED}(Mbps){YELLOW} > {DarkBLUE}"))
+                    size = float(input(f"{YELLOW}your file size {RED}(Mbit){YELLOW} > {DarkBLUE}"))
                     break
                 except ValueError:
                     print(f"{RED}invalid input.")
             calculate_time(size,speed)
             break
 
-        elif user_input == "2":
+        elif user_input == "2":         # means the user chose to calculate the speed of the download
             while True:
                 try:
-                    time = int(input(f"{YELLOW}the time {RED}(seconds){YELLOW} > {DarkBLUE}"))
-                    size = int(input(f"{YELLOW}your file size {RED}(Mbit){YELLOW} > {DarkBLUE}"))
+                    time = float(input(f"{YELLOW}the time {RED}(seconds){YELLOW} > {DarkBLUE}"))
+                    size = float(input(f"{YELLOW}your file size {RED}(Mbit){YELLOW} > {DarkBLUE}"))
                     break
                 except ValueError:
                     print(f"{RED}invalid input.")
             calculate_speed(size, time)
             break
         
-        elif user_input == "3":
+        elif user_input == "3":         # means the user chose to calculate the file size that was downloaded
             while True:
                 try:
-                    time = int(input(f"{YELLOW}the time {RED}(seconds){YELLOW} > {DarkBLUE}"))
-                    speed = int(input(f"{YELLOW}your internet speed {RED}(Mbps){YELLOW} > {DarkBLUE}"))
+                    time = float(input(f"{YELLOW}the time {RED}(seconds){YELLOW} > {DarkBLUE}"))
+                    speed = float(input(f"{YELLOW}your internet speed {RED}(Mbps){YELLOW} > {DarkBLUE}"))
                     break
                 except ValueError:
                     print(f"{RED}invalid input.")
             calculate_fileSize(time, speed)
             break
 
-        elif user_input.lower() == "help" or user_input.lower() == "h":
+        elif user_input.lower() == "help" or user_input.lower() == "h":     # to help the user
             print(f'''
             {PURPLE}    [1]{GREY} to calculate the time
             {PURPLE}    [2]{GREY} to calculate the speed
@@ -60,14 +60,14 @@ while True: #main loop
             {PURPLE}        [Q]{GREY} to exit the code
             ''')
         
-        elif user_input.lower() == "cls" or user_input.lower() == "clear":
+        elif user_input.lower() == "cls" or user_input.lower() == "clear":  # to clear the the text
             os.system("cls")
         
-        elif user_input.lower() == "q" or user_input.lower() == "exit":
+        elif user_input.lower() == "q" or user_input.lower() == "exit":     # to exit the program
             print(GREY)
             exit()
         
-        else:
+        else:                                                               # the user messed up
             print(f"{RED}invalid input, please choose one of the choices above")
 
     print(f"{GREY}\banything else?(type \"help\" to display the all the functions)")
